+++ src/database/migrations/20200910143718_create_table_articles.js
@@ -1,0 +1,11 @@
+exports.up = async knex => knex.schema.createTable('articles', table => {
+    table.increments('id');
+    table.text('title').unique().notNullable();
+    table.text('body').notNullable();
+    table.text('slug').notNullable();
+
+    table.timestamp('created_at').defaultTo(knex.fn.now());
+    table.timestamp('updated_at').defaultTo(knex.fn.now());
+});
+
+exports.down = async knex => knex.schema.dropTable('articles');