+++ src/server.js
@@ -1,12 +1,12 @@
 require('dotenv/config');
 
 const express = require('express');
+const knex = require('./database');
 const app = express();
 
-app.get('/', (req,res) => {
-    res.send('Hello World!');
+app.get('/', async (req,res) => {
+    const results = await knex('articles');
+    res.json(results);
 });
 
-app.listen(process.env.PORT || 3001, () => {
-    console.log('Server is running')
-});
+app.listen(process.env.PORT || 3001);