+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change log
 
+## v0.2.0 - Sep. 10, 2020
+
+Connection with database (postgres)
+
+**News features:**
+- Dataabase connection
+
+**Added:**
+- [LICENSE.md](LICENSE.md)
+- knexfile.js
+
 ## v0.1.0 - Sep. 10, 2020
 
 Initial release.
@@ -8,9 +19,8 @@ Initial release.
 - Hello World
 
 **Added:**
-- README.md
-- CHANGELOG.md
+- [README.md](README.md)
+- [CHANGELOG.md](CHANGELOG.md)
 - .env.example
 - src/server.js
 - src/database/index.js
-- src/config/database.js